% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{allocate_static_ip}
\alias{allocate_static_ip}
\title{Allocates a static IP address}
\usage{
allocate_static_ip(staticIpName)
}
\arguments{
\item{staticIpName}{The name of the static IP address.}
}
\description{
Allocates a static IP address.
}
\examples{

}
