% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{attach_instances_to_load_balancer}
\alias{attach_instances_to_load_balancer}
\title{Attaches one or more Lightsail instances to a load balancer}
\usage{
attach_instances_to_load_balancer(loadBalancerName, instanceNames)
}
\arguments{
\item{loadBalancerName}{The name of the load balancer.}

\item{instanceNames}{An array of strings representing the instance name(s) you want to attach to your load balancer.

An instance must be \code{running} before you can attach it to your load balancer.

There are no additional limits on the number of instances you can attach to your load balancer, aside from the limit of Lightsail instances you can create in your account (20).}
}
\description{
Attaches one or more Lightsail instances to a load balancer.
}
\details{
After some time, the instances are attached to the load balancer and the health check status is available.
}
\examples{

}
