% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{attach_load_balancer_tls_certificate}
\alias{attach_load_balancer_tls_certificate}
\title{Attaches a Transport Layer Security (TLS) certificate to your load balancer}
\usage{
attach_load_balancer_tls_certificate(loadBalancerName, certificateName)
}
\arguments{
\item{loadBalancerName}{The name of the load balancer to which you want to associate the SSL/TLS certificate.}

\item{certificateName}{The name of your SSL/TLS certificate.}
}
\description{
Attaches a Transport Layer Security (TLS) certificate to your load balancer. TLS is just an updated, more secure version of Secure Socket Layer (SSL).
}
\details{
Once you create and validate your certificate, you can attach it to your load balancer. You can also use this API to rotate the certificates on your account. Use the \code{AttachLoadBalancerTlsCertificate} operation with the non-attached certificate, and it will replace the existing one and become the attached certificate.
}
\examples{

}
