% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{attach_static_ip}
\alias{attach_static_ip}
\title{Attaches a static IP address to a specific Amazon Lightsail instance}
\usage{
attach_static_ip(staticIpName, instanceName)
}
\arguments{
\item{staticIpName}{The name of the static IP.}

\item{instanceName}{The instance name to which you want to attach the static IP address.}
}
\description{
Attaches a static IP address to a specific Amazon Lightsail instance.
}
\examples{

}
