% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{close_instance_public_ports}
\alias{close_instance_public_ports}
\title{Closes the public ports on a specific Amazon Lightsail instance}
\usage{
close_instance_public_ports(portInfo, instanceName)
}
\arguments{
\item{portInfo}{Information about the public port you are trying to close.}

\item{instanceName}{The name of the instance on which you're attempting to close the public ports.}
}
\description{
Closes the public ports on a specific Amazon Lightsail instance.
}
\examples{

}
