% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{create_disk}
\alias{create_disk}
\title{Creates a block storage disk that can be attached to a Lightsail instance in the same Availability Zone (e}
\usage{
create_disk(diskName, availabilityZone, sizeInGb)
}
\arguments{
\item{diskName}{The unique Lightsail disk name (e.g., \code{my-disk}).}

\item{availabilityZone}{The Availability Zone where you want to create the disk (e.g., \code{us-east-2a}). Choose the same Availability Zone as the Lightsail instance where you want to create the disk.

Use the GetRegions operation to list the Availability Zones where Lightsail is currently available.}

\item{sizeInGb}{The size of the disk in GB (e.g., \code{32}).}
}
\description{
Creates a block storage disk that can be attached to a Lightsail instance in the same Availability Zone (e.g., \code{us-east-2a}). The disk is created in the regional endpoint that you send the HTTP request to. For more information, see \href{https://lightsail.aws.amazon.com/ls/docs/overview/article/understanding-regions-and-availability-zones-in-amazon-lightsail}{Regions and Availability Zones in Lightsail}.
}
\examples{

}
