% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{create_domain_entry}
\alias{create_domain_entry}
\title{Creates one of the following entry records associated with the domain: A record, CNAME record, TXT record, or MX record}
\usage{
create_domain_entry(domainName, domainEntry)
}
\arguments{
\item{domainName}{The domain name (e.g., \code{example.com}) for which you want to create the domain entry.}

\item{domainEntry}{An array of key-value pairs containing information about the domain entry request.}
}
\description{
Creates one of the following entry records associated with the domain: A record, CNAME record, TXT record, or MX record.
}
\examples{

}
