% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{create_instance_snapshot}
\alias{create_instance_snapshot}
\title{Creates a snapshot of a specific virtual private server, or \emph{instance}}
\usage{
create_instance_snapshot(instanceSnapshotName, instanceName)
}
\arguments{
\item{instanceSnapshotName}{The name for your new snapshot.}

\item{instanceName}{The Lightsail instance on which to base your snapshot.}
}
\description{
Creates a snapshot of a specific virtual private server, or \emph{instance}. You can use a snapshot to create a new instance that is based on that snapshot.
}
\examples{

}
