% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{create_instances_from_snapshot}
\alias{create_instances_from_snapshot}
\title{Uses a specific snapshot as a blueprint for creating one or more new instances that are based on that identical configuration}
\usage{
create_instances_from_snapshot(instanceNames, availabilityZone,
  instanceSnapshotName, bundleId, attachedDiskMapping = NULL,
  userData = NULL, keyPairName = NULL)
}
\arguments{
\item{instanceNames}{The names for your new instances.}

\item{availabilityZone}{The Availability Zone where you want to create your instances. Use the following formatting: \code{us-east-2a} (case sensitive). You can get a list of availability zones by using the \href{http://docs.aws.amazon.com/lightsail/2016-11-28/api-reference/API_GetRegions.html}{get regions} operation. Be sure to add the \code{include availability zones} parameter to your request.}

\item{instanceSnapshotName}{The name of the instance snapshot on which you are basing your new instances. Use the get instance snapshots operation to return information about your existing snapshots.}

\item{bundleId}{The bundle of specification information for your virtual private server (or \emph{instance}), including the pricing plan (e.g., \code{micro_1_0}).}

\item{attachedDiskMapping}{An object containing information about one or more disk mappings.}

\item{userData}{You can create a launch script that configures a server with additional user data. For example, \code{apt-get -y update}.

Depending on the machine image you choose, the command to get software on your instance varies. Amazon Linux and CentOS use \code{yum}, Debian and Ubuntu use \code{apt-get}, and FreeBSD uses \code{pkg}. For a complete list, see the \href{http://lightsail.aws.amazon.com/ls/docs/getting-started/articles/pre-installed-apps}{Dev Guide}.}

\item{keyPairName}{The name for your key pair.}
}
\description{
Uses a specific snapshot as a blueprint for creating one or more new instances that are based on that identical configuration.
}
\examples{

}
