% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{create_key_pair}
\alias{create_key_pair}
\title{Creates an SSH key pair}
\usage{
create_key_pair(keyPairName)
}
\arguments{
\item{keyPairName}{[required] The name for your new key pair.}
}
\description{
Creates an SSH key pair.
}
\section{Accepted Parameters}{
\preformatted{create_key_pair(
  keyPairName = "string"
)
}
}

