% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{delete_disk}
\alias{delete_disk}
\title{Deletes the specified block storage disk}
\usage{
delete_disk(diskName)
}
\arguments{
\item{diskName}{[required] The unique name of the disk you want to delete (e.g., \code{my-disk}).}
}
\description{
Deletes the specified block storage disk. The disk must be in the \code{available} state (not attached to a Lightsail instance).
}
\details{
The disk may remain in the \code{deleting} state for several minutes.
}
\section{Accepted Parameters}{
\preformatted{delete_disk(
  diskName = "string"
)
}
}

