% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{delete_domain}
\alias{delete_domain}
\title{Deletes the specified domain recordset and all of its domain records}
\usage{
delete_domain(domainName)
}
\arguments{
\item{domainName}{[required] The specific domain name to delete.}
}
\description{
Deletes the specified domain recordset and all of its domain records.
}
\details{
The \code{delete domain} operation supports tag-based access control via resource tags applied to the resource identified by domainName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\section{Accepted Parameters}{
\preformatted{delete_domain(
  domainName = "string"
)
}
}

