% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{delete_domain_entry}
\alias{delete_domain_entry}
\title{Deletes a specific domain entry}
\usage{
delete_domain_entry(domainName, domainEntry)
}
\arguments{
\item{domainName}{The name of the domain entry to delete.}

\item{domainEntry}{An array of key-value pairs containing information about your domain entries.}
}
\description{
Deletes a specific domain entry.
}
\examples{

}
