% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{delete_instance_snapshot}
\alias{delete_instance_snapshot}
\title{Deletes a specific snapshot of a virtual private server (or \emph{instance})}
\usage{
delete_instance_snapshot(instanceSnapshotName)
}
\arguments{
\item{instanceSnapshotName}{[required] The name of the snapshot to delete.}
}
\description{
Deletes a specific snapshot of a virtual private server (or \emph{instance}).
}
\section{Accepted Parameters}{
\preformatted{delete_instance_snapshot(
  instanceSnapshotName = "string"
)
}
}

