% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{delete_key_pair}
\alias{delete_key_pair}
\title{Deletes a specific SSH key pair}
\usage{
delete_key_pair(keyPairName)
}
\arguments{
\item{keyPairName}{[required] The name of the key pair to delete.}
}
\description{
Deletes a specific SSH key pair.
}
\section{Accepted Parameters}{
\preformatted{delete_key_pair(
  keyPairName = "string"
)
}
}

