% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{delete_load_balancer_tls_certificate}
\alias{delete_load_balancer_tls_certificate}
\title{Deletes an SSL/TLS certificate associated with a Lightsail load balancer}
\usage{
delete_load_balancer_tls_certificate(loadBalancerName, certificateName,
  force = NULL)
}
\arguments{
\item{loadBalancerName}{The load balancer name.}

\item{certificateName}{The SSL/TLS certificate name.}

\item{force}{When \code{true}, forces the deletion of an SSL/TLS certificate.

There can be two certificates associated with a Lightsail load balancer: the primary and the backup. The force parameter is required when the primary SSL/TLS certificate is in use by an instance attached to the load balancer.}
}
\description{
Deletes an SSL/TLS certificate associated with a Lightsail load balancer.
}
\examples{

}
