% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{delete_relational_database_snapshot}
\alias{delete_relational_database_snapshot}
\title{Deletes a database snapshot in Amazon Lightsail}
\usage{
delete_relational_database_snapshot(relationalDatabaseSnapshotName)
}
\arguments{
\item{relationalDatabaseSnapshotName}{[required] The name of the database snapshot that you are deleting.}
}
\description{
Deletes a database snapshot in Amazon Lightsail.
}
\section{Accepted Parameters}{
\preformatted{delete_relational_database_snapshot(
  relationalDatabaseSnapshotName = "string"
)
}
}

