% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{detach_static_ip}
\alias{detach_static_ip}
\title{Detaches a static IP from the Amazon Lightsail instance to which it is attached}
\usage{
detach_static_ip(staticIpName)
}
\arguments{
\item{staticIpName}{The name of the static IP to detach from the instance.}
}
\description{
Detaches a static IP from the Amazon Lightsail instance to which it is attached.
}
\examples{

}
