% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{download_default_key_pair}
\alias{download_default_key_pair}
\title{Downloads the default SSH key pair from the user's account}
\usage{
download_default_key_pair()
}
\description{
Downloads the default SSH key pair from the user's account.
}
\examples{

}
