% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_active_names}
\alias{get_active_names}
\title{Returns the names of all active (not deleted) resources}
\usage{
get_active_names(pageToken = NULL)
}
\arguments{
\item{pageToken}{A token used for paginating results from your get active names request.}
}
\description{
Returns the names of all active (not deleted) resources.
}
\examples{

}
