% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_blueprints}
\alias{get_blueprints}
\title{Returns the list of available instance images, or \emph{blueprints}}
\usage{
get_blueprints(
  includeInactive = TRUE|FALSE,
  pageToken = "string"
)
}
\arguments{
\item{includeInactive}{A Boolean value indicating whether to include inactive results in your request.}

\item{pageToken}{A token used for advancing to the next page of results from your get blueprints request.}
}
\description{
Returns the list of available instance images, or \emph{blueprints}. You can use a blueprint to create a new virtual private server already running a specific operating system, as well as a preinstalled app or development stack. The software each instance is running depends on the blueprint image you choose.
}
\examples{

}
