% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_bundles}
\alias{get_bundles}
\title{Returns the list of bundles that are available for purchase}
\usage{
get_bundles(includeInactive = NULL, pageToken = NULL)
}
\arguments{
\item{includeInactive}{A Boolean value that indicates whether to include inactive bundle results in your request.}

\item{pageToken}{A token used for advancing to the next page of results from your get bundles request.}
}
\description{
Returns the list of bundles that are available for purchase. A bundle describes the specs for your virtual private server (or \emph{instance}).
}
\examples{

}
