% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_disk}
\alias{get_disk}
\title{Returns information about a specific block storage disk}
\usage{
get_disk(diskName)
}
\arguments{
\item{diskName}{The name of the disk (e.g., \code{my-disk}).}
}
\description{
Returns information about a specific block storage disk.
}
\examples{

}
