% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_disk_snapshot}
\alias{get_disk_snapshot}
\title{Returns information about a specific block storage disk snapshot}
\usage{
get_disk_snapshot(diskSnapshotName)
}
\arguments{
\item{diskSnapshotName}{[required] The name of the disk snapshot (e.g., \code{my-disk-snapshot}).}
}
\description{
Returns information about a specific block storage disk snapshot.
}
\section{Accepted Parameters}{
\preformatted{get_disk_snapshot(
  diskSnapshotName = "string"
)
}
}

