% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_domain}
\alias{get_domain}
\title{Returns information about a specific domain recordset}
\usage{
get_domain(domainName)
}
\arguments{
\item{domainName}{The domain name for which your want to return information about.}
}
\description{
Returns information about a specific domain recordset.
}
\examples{

}
