% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_domains}
\alias{get_domains}
\title{Returns a list of all domains in the user's account}
\usage{
get_domains(
  pageToken = "string"
)
}
\arguments{
\item{pageToken}{A token used for advancing to the next page of results from your get domains request.}
}
\description{
Returns a list of all domains in the user's account.
}
\examples{

}
