% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_instance}
\alias{get_instance}
\title{Returns information about a specific Amazon Lightsail instance, which is a virtual private server}
\usage{
get_instance(
  instanceName = "string"
)
}
\arguments{
\item{instanceName}{[required] The name of the instance.}
}
\description{
Returns information about a specific Amazon Lightsail instance, which is a virtual private server.
}
\examples{

}
