% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_instance_metric_data}
\alias{get_instance_metric_data}
\title{Returns the data points for the specified Amazon Lightsail instance metric, given an instance name}
\usage{
get_instance_metric_data(instanceName, metricName, period, startTime, endTime,
  unit, statistics)
}
\arguments{
\item{instanceName}{The name of the instance for which you want to get metrics data.}

\item{metricName}{The metric name to get data about.}

\item{period}{The time period for which you are requesting data.}

\item{startTime}{The start time of the time period.}

\item{endTime}{The end time of the time period.}

\item{unit}{The unit. The list of valid values is below.}

\item{statistics}{The instance statistics.}
}
\description{
Returns the data points for the specified Amazon Lightsail instance metric, given an instance name.
}
\examples{

}
