% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_instance_port_states}
\alias{get_instance_port_states}
\title{Returns the port states for a specific virtual private server, or \emph{instance}}
\usage{
get_instance_port_states(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the instance.}
}
\description{
Returns the port states for a specific virtual private server, or \emph{instance}.
}
\section{Accepted Parameters}{
\preformatted{get_instance_port_states(
  instanceName = "string"
)
}
}

