% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_instance_snapshot}
\alias{get_instance_snapshot}
\title{Returns information about a specific instance snapshot}
\usage{
get_instance_snapshot(instanceSnapshotName)
}
\arguments{
\item{instanceSnapshotName}{The name of the snapshot for which you are requesting information.}
}
\description{
Returns information about a specific instance snapshot.
}
\examples{

}
