% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_instance_snapshots}
\alias{get_instance_snapshots}
\title{Returns all instance snapshots for the user's account}
\usage{
get_instance_snapshots(
  pageToken = "string"
)
}
\arguments{
\item{pageToken}{A token used for advancing to the next page of results from your get instance snapshots request.}
}
\description{
Returns all instance snapshots for the user's account.
}
\examples{

}
