% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_instances}
\alias{get_instances}
\title{Returns information about all Amazon Lightsail virtual private servers, or \emph{instances}}
\usage{
get_instances(pageToken = NULL)
}
\arguments{
\item{pageToken}{A token used for advancing to the next page of results from your get instances request.}
}
\description{
Returns information about all Amazon Lightsail virtual private servers, or \emph{instances}.
}
\section{Accepted Parameters}{
\preformatted{get_instances(
  pageToken = "string"
)
}
}

