% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_key_pair}
\alias{get_key_pair}
\title{Returns information about a specific key pair}
\usage{
get_key_pair(keyPairName)
}
\arguments{
\item{keyPairName}{[required] The name of the key pair for which you are requesting information.}
}
\description{
Returns information about a specific key pair.
}
\section{Accepted Parameters}{
\preformatted{get_key_pair(
  keyPairName = "string"
)
}
}

