% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_key_pairs}
\alias{get_key_pairs}
\title{Returns information about all key pairs in the user's account}
\usage{
get_key_pairs(pageToken = NULL)
}
\arguments{
\item{pageToken}{A token used for advancing to the next page of results from your get key pairs request.}
}
\description{
Returns information about all key pairs in the user's account.
}
\section{Accepted Parameters}{
\preformatted{get_key_pairs(
  pageToken = "string"
)
}
}

