% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_load_balancer}
\alias{get_load_balancer}
\title{Returns information about the specified Lightsail load balancer}
\usage{
get_load_balancer(
  loadBalancerName = "string"
)
}
\arguments{
\item{loadBalancerName}{[required] The name of the load balancer.}
}
\description{
Returns information about the specified Lightsail load balancer.
}
\examples{

}
