% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_load_balancer_tls_certificates}
\alias{get_load_balancer_tls_certificates}
\title{Returns information about the TLS certificates that are associated with the specified Lightsail load balancer}
\usage{
get_load_balancer_tls_certificates(
  loadBalancerName = "string"
)
}
\arguments{
\item{loadBalancerName}{[required] The name of the load balancer you associated with your SSL/TLS certificate.}
}
\description{
Returns information about the TLS certificates that are associated with the specified Lightsail load balancer.
}
\details{
TLS is just an updated, more secure version of Secure Socket Layer (SSL).

You can have a maximum of 2 certificates associated with a Lightsail load balancer. One is active and the other is inactive.
}
\examples{

}
