% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_load_balancers}
\alias{get_load_balancers}
\title{Returns information about all load balancers in an account}
\usage{
get_load_balancers(pageToken = NULL)
}
\arguments{
\item{pageToken}{A token used for paginating the results from your GetLoadBalancers request.}
}
\description{
Returns information about all load balancers in an account.
}
\details{
If you are describing a long list of load balancers, you can paginate the output to make the list more manageable. You can use the pageToken and nextPageToken values to retrieve the next items in the list.
}
\section{Accepted Parameters}{
\preformatted{get_load_balancers(
  pageToken = "string"
)
}
}

