% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_operation}
\alias{get_operation}
\title{Returns information about a specific operation}
\usage{
get_operation(operationId)
}
\arguments{
\item{operationId}{[required] A GUID used to identify the operation.}
}
\description{
Returns information about a specific operation. Operations include events such as when you create an instance, allocate a static IP, attach a static IP, and so on.
}
\section{Accepted Parameters}{
\preformatted{get_operation(
  operationId = "string"
)
}
}

