% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_operations}
\alias{get_operations}
\title{Returns information about all operations}
\usage{
get_operations(pageToken = NULL)
}
\arguments{
\item{pageToken}{A token used for advancing to the next page of results from your get operations request.}
}
\description{
Returns information about all operations.
}
\details{
Results are returned from oldest to newest, up to a maximum of 200. Results can be paged by making each subsequent call to \code{GetOperations} use the maximum (last) \code{statusChangedAt} value from the previous request.
}
\section{Accepted Parameters}{
\preformatted{get_operations(
  pageToken = "string"
)
}
}

