% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_relational_database}
\alias{get_relational_database}
\title{Returns information about a specific database in Amazon Lightsail}
\usage{
get_relational_database(
  relationalDatabaseName = "string"
)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of the database that you are looking up.}
}
\description{
Returns information about a specific database in Amazon Lightsail.
}
\examples{

}
