% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_relational_database_blueprints}
\alias{get_relational_database_blueprints}
\title{Returns a list of available database blueprints in Amazon Lightsail}
\usage{
get_relational_database_blueprints(pageToken = NULL)
}
\arguments{
\item{pageToken}{A token used for advancing to a specific page of results for your \code{get relational database blueprints} request.}
}
\description{
Returns a list of available database blueprints in Amazon Lightsail. A blueprint describes the major engine version of a database.
}
\details{
You can use a blueprint ID to create a new database that runs a specific database engine.
}
\section{Accepted Parameters}{
\preformatted{get_relational_database_blueprints(
  pageToken = "string"
)
}
}

