% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_relational_database_log_streams}
\alias{get_relational_database_log_streams}
\title{Returns a list of available log streams for a specific database in Amazon Lightsail}
\usage{
get_relational_database_log_streams(
  relationalDatabaseName = "string"
)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database for which to get log streams.}
}
\description{
Returns a list of available log streams for a specific database in Amazon Lightsail.
}
\examples{

}
