% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_relational_database_parameters}
\alias{get_relational_database_parameters}
\title{Returns all of the runtime parameters offered by the underlying database software, or engine, for a specific database in Amazon Lightsail}
\usage{
get_relational_database_parameters(
  relationalDatabaseName = "string",
  pageToken = "string"
)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database for which to get parameters.}

\item{pageToken}{A token used for advancing to a specific page of results for your \code{get relational database parameters} request.}
}
\description{
Returns all of the runtime parameters offered by the underlying database software, or engine, for a specific database in Amazon Lightsail.
}
\details{
In addition to the parameter names and values, this operation returns other information about each parameter. This information includes whether changes require a reboot, whether the parameter is modifiable, the allowed values, and the data types.
}
\examples{

}
