% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_relational_database_snapshot}
\alias{get_relational_database_snapshot}
\title{Returns information about a specific database snapshot in Amazon Lightsail}
\usage{
get_relational_database_snapshot(
  relationalDatabaseSnapshotName = "string"
)
}
\arguments{
\item{relationalDatabaseSnapshotName}{[required] The name of the database snapshot for which to get information.}
}
\description{
Returns information about a specific database snapshot in Amazon Lightsail.
}
\examples{

}
