% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_relational_database_snapshots}
\alias{get_relational_database_snapshots}
\title{Returns information about all of your database snapshots in Amazon Lightsail}
\usage{
get_relational_database_snapshots(
  pageToken = "string"
)
}
\arguments{
\item{pageToken}{A token used for advancing to a specific page of results for your \code{get relational database snapshots} request.}
}
\description{
Returns information about all of your database snapshots in Amazon Lightsail.
}
\examples{

}
