% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_relational_databases}
\alias{get_relational_databases}
\title{Returns information about all of your databases in Amazon Lightsail}
\usage{
get_relational_databases(pageToken = NULL)
}
\arguments{
\item{pageToken}{A token used for advancing to a specific page of results for your get relational database request.}
}
\description{
Returns information about all of your databases in Amazon Lightsail.
}
\section{Accepted Parameters}{
\preformatted{get_relational_databases(
  pageToken = "string"
)
}
}

