% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_static_ip}
\alias{get_static_ip}
\title{Returns information about a specific static IP}
\usage{
get_static_ip(staticIpName)
}
\arguments{
\item{staticIpName}{[required] The name of the static IP in Lightsail.}
}
\description{
Returns information about a specific static IP.
}
\section{Accepted Parameters}{
\preformatted{get_static_ip(
  staticIpName = "string"
)
}
}

