% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_static_ips}
\alias{get_static_ips}
\title{Returns information about all static IPs in the user's account}
\usage{
get_static_ips(pageToken = NULL)
}
\arguments{
\item{pageToken}{A token used for advancing to the next page of results from your get static IPs request.}
}
\description{
Returns information about all static IPs in the user's account.
}
\examples{

}
