% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{import_key_pair}
\alias{import_key_pair}
\title{Imports a public SSH key from a specific key pair}
\usage{
import_key_pair(keyPairName, publicKeyBase64)
}
\arguments{
\item{keyPairName}{The name of the key pair for which you want to import the public key.}

\item{publicKeyBase64}{A base64-encoded public key of the \code{ssh-rsa} type.}
}
\description{
Imports a public SSH key from a specific key pair.
}
\examples{

}
