% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{is_vpc_peered}
\alias{is_vpc_peered}
\title{Returns a Boolean value indicating whether your Lightsail VPC is peered}
\usage{
is_vpc_peered()
}
\description{
Returns a Boolean value indicating whether your Lightsail VPC is peered.
}
\section{Accepted Parameters}{
\preformatted{is_vpc_peered()
}
}

