% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{open_instance_public_ports}
\alias{open_instance_public_ports}
\title{Adds public ports to an Amazon Lightsail instance}
\usage{
open_instance_public_ports(portInfo, instanceName)
}
\arguments{
\item{portInfo}{An array of key-value pairs containing information about the port mappings.}

\item{instanceName}{The name of the instance for which you want to open the public ports.}
}
\description{
Adds public ports to an Amazon Lightsail instance.
}
\examples{

}
