% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_package.R
\docType{package}
\name{paws.lightsail-package}
\alias{paws.lightsail}
\alias{paws.lightsail-package}
\title{paws.lightsail: Amazon Lightsail}
\description{
Amazon Lightsail is the easiest way to get started
with AWS for developers who just need virtual private servers.
Lightsail includes everything you need to launch your project quickly
- a virtual machine, SSD-based storage, data transfer, DNS management,
and a static IP - for a low, predictable price. You manage those
Lightsail servers through the Lightsail console or by using the API or
command-line interface (CLI).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
