% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{peer_vpc}
\alias{peer_vpc}
\title{Tries to peer the Lightsail VPC with the user's default VPC}
\usage{
peer_vpc()
}
\description{
Tries to peer the Lightsail VPC with the user's default VPC.
}
\section{Accepted Parameters}{
\preformatted{peer_vpc()
}
}

