% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{put_instance_public_ports}
\alias{put_instance_public_ports}
\title{Sets the specified open ports for an Amazon Lightsail instance, and closes all ports for every protocol not included in the current request}
\usage{
put_instance_public_ports(portInfos, instanceName)
}
\arguments{
\item{portInfos}{Specifies information about the public port(s).}

\item{instanceName}{The Lightsail instance name of the public port(s) you are setting.}
}
\description{
Sets the specified open ports for an Amazon Lightsail instance, and closes all ports for every protocol not included in the current request.
}
\examples{

}
