% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{reboot_relational_database}
\alias{reboot_relational_database}
\title{Restarts a specific database in Amazon Lightsail}
\usage{
reboot_relational_database(relationalDatabaseName)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database to reboot.}
}
\description{
Restarts a specific database in Amazon Lightsail.
}
\section{Accepted Parameters}{
\preformatted{reboot_relational_database(
  relationalDatabaseName = "string"
)
}
}

