% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{release_static_ip}
\alias{release_static_ip}
\title{Deletes a specific static IP from your account}
\usage{
release_static_ip(staticIpName)
}
\arguments{
\item{staticIpName}{[required] The name of the static IP to delete.}
}
\description{
Deletes a specific static IP from your account.
}
\section{Accepted Parameters}{
\preformatted{release_static_ip(
  staticIpName = "string"
)
}
}

