% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{stop_instance}
\alias{stop_instance}
\title{Stops a specific Amazon Lightsail instance that is currently running}
\usage{
stop_instance(instanceName, force = NULL)
}
\arguments{
\item{instanceName}{[required] The name of the instance (a virtual private server) to stop.}

\item{force}{When set to \code{True}, forces a Lightsail instance that is stuck in a \code{stopping} state to stop.

Only use the \code{force} parameter if your instance is stuck in the \code{stopping} state. In any other state, your instance should stop normally without adding this parameter to your API request.}
}
\description{
Stops a specific Amazon Lightsail instance that is currently running.
}
\section{Accepted Parameters}{
\preformatted{stop_instance(
  instanceName = "string",
  force = TRUE|FALSE
)
}
}

