% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{stop_relational_database}
\alias{stop_relational_database}
\title{Stops a specific database that is currently running in Amazon Lightsail}
\usage{
stop_relational_database(relationalDatabaseName,
  relationalDatabaseSnapshotName = NULL)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database to stop.}

\item{relationalDatabaseSnapshotName}{The name of your new database snapshot to be created before stopping your database.}
}
\description{
Stops a specific database that is currently running in Amazon Lightsail.
}
\section{Accepted Parameters}{
\preformatted{stop_relational_database(
  relationalDatabaseName = "string",
  relationalDatabaseSnapshotName = "string"
)
}
}

