% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{attach_disk}
\alias{attach_disk}
\title{Attaches a block storage disk to a running or stopped Lightsail instance and exposes it to the instance with the specified disk name}
\usage{
attach_disk(
  diskName = "string",
  instanceName = "string",
  diskPath = "string"
)
}
\arguments{
\item{diskName}{[required] The unique Lightsail disk name (e.g., \code{my-disk}).}

\item{instanceName}{[required] The name of the Lightsail instance where you want to utilize the storage disk.}

\item{diskPath}{[required] The disk path to expose to the instance (e.g., \code{/dev/xvdf}).}
}
\description{
Attaches a block storage disk to a running or stopped Lightsail instance and exposes it to the instance with the specified disk name.
}
\details{
The \code{attach disk} operation supports tag-based access control via resource tags applied to the resource identified by diskName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\examples{

}
