% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{create_domain}
\alias{create_domain}
\title{Creates a domain resource for the specified domain (e}
\usage{
create_domain(domainName)
}
\arguments{
\item{domainName}{The domain name to manage (e.g., \code{example.com}).

You cannot register a new domain name using Lightsail. You must register a domain name using Amazon Route 53 or another domain name registrar. If you have already registered your domain, you can enter its name in this parameter to manage the DNS records for that domain.}
}
\description{
Creates a domain resource for the specified domain (e.g., example.com).
}
\examples{

}
