% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{create_key_pair}
\alias{create_key_pair}
\title{Creates sn SSH key pair}
\usage{
create_key_pair(keyPairName)
}
\arguments{
\item{keyPairName}{The name for your new key pair.}
}
\description{
Creates sn SSH key pair.
}
\examples{

}
