% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{delete_domain}
\alias{delete_domain}
\title{Deletes the specified domain recordset and all of its domain records}
\usage{
delete_domain(domainName)
}
\arguments{
\item{domainName}{The specific domain name to delete.}
}
\description{
Deletes the specified domain recordset and all of its domain records.
}
\examples{

}
