% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{delete_instance_snapshot}
\alias{delete_instance_snapshot}
\title{Deletes a specific snapshot of a virtual private server (or \emph{instance})}
\usage{
delete_instance_snapshot(
  instanceSnapshotName = "string"
)
}
\arguments{
\item{instanceSnapshotName}{[required] The name of the snapshot to delete.}
}
\description{
Deletes a specific snapshot of a virtual private server (or \emph{instance}).
}
\details{
The \code{delete instance snapshot} operation supports tag-based access control via resource tags applied to the resource identified by instanceSnapshotName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\examples{

}
