% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{delete_load_balancer}
\alias{delete_load_balancer}
\title{Deletes a Lightsail load balancer and all its associated SSL/TLS certificates}
\usage{
delete_load_balancer(
  loadBalancerName = "string"
)
}
\arguments{
\item{loadBalancerName}{[required] The name of the load balancer you want to delete.}
}
\description{
Deletes a Lightsail load balancer and all its associated SSL/TLS certificates. Once the load balancer is deleted, you will need to create a new load balancer, create a new certificate, and verify domain ownership again.
}
\details{
The \code{delete load balancer} operation supports tag-based access control via resource tags applied to the resource identified by loadBalancerName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\examples{

}
