% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{delete_load_balancer_tls_certificate}
\alias{delete_load_balancer_tls_certificate}
\title{Deletes an SSL/TLS certificate associated with a Lightsail load balancer}
\usage{
delete_load_balancer_tls_certificate(
  loadBalancerName = "string",
  certificateName = "string",
  force = TRUE|FALSE
)
}
\arguments{
\item{loadBalancerName}{[required] The load balancer name.}

\item{certificateName}{[required] The SSL/TLS certificate name.}

\item{force}{When \code{true}, forces the deletion of an SSL/TLS certificate.

There can be two certificates associated with a Lightsail load balancer: the primary and the backup. The \code{force} parameter is required when the primary SSL/TLS certificate is in use by an instance attached to the load balancer.}
}
\description{
Deletes an SSL/TLS certificate associated with a Lightsail load balancer.
}
\details{
The \code{delete load balancer tls certificate} operation supports tag-based access control via resource tags applied to the resource identified by loadBalancerName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\examples{

}
