% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{detach_instances_from_load_balancer}
\alias{detach_instances_from_load_balancer}
\title{Detaches the specified instances from a Lightsail load balancer}
\usage{
detach_instances_from_load_balancer(loadBalancerName, instanceNames)
}
\arguments{
\item{loadBalancerName}{[required] The name of the Lightsail load balancer.}

\item{instanceNames}{[required] An array of strings containing the names of the instances you want to detach from the load balancer.}
}
\description{
Detaches the specified instances from a Lightsail load balancer.
}
\details{
This operation waits until the instances are no longer needed before they are detached from the load balancer.
}
\section{Accepted Parameters}{
\preformatted{detach_instances_from_load_balancer(
  loadBalancerName = "string",
  instanceNames = list(
    "string"
  )
)
}
}

