% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{get_instance_access_details}
\alias{get_instance_access_details}
\title{Returns temporary SSH keys you can use to connect to a specific virtual private server, or \emph{instance}}
\usage{
get_instance_access_details(instanceName, protocol = NULL)
}
\arguments{
\item{instanceName}{The name of the instance to access.}

\item{protocol}{The protocol to use to connect to your instance. Defaults to \code{ssh}.}
}
\description{
Returns temporary SSH keys you can use to connect to a specific virtual private server, or \emph{instance}.
}
\examples{

}
