% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{reboot_instance}
\alias{reboot_instance}
\title{Restarts a specific instance}
\usage{
reboot_instance(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the instance to reboot.}
}
\description{
Restarts a specific instance.
}
\details{
The \code{reboot instance} operation supports tag-based access control via resource tags applied to the resource identified by instanceName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\section{Accepted Parameters}{
\preformatted{reboot_instance(
  instanceName = "string"
)
}
}

