% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{reboot_relational_database}
\alias{reboot_relational_database}
\title{Restarts a specific database in Amazon Lightsail}
\usage{
reboot_relational_database(relationalDatabaseName)
}
\arguments{
\item{relationalDatabaseName}{[required] The name of your database to reboot.}
}
\description{
Restarts a specific database in Amazon Lightsail.
}
\details{
The \code{reboot relational database} operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\section{Accepted Parameters}{
\preformatted{reboot_relational_database(
  relationalDatabaseName = "string"
)
}
}

