% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{start_instance}
\alias{start_instance}
\title{Starts a specific Amazon Lightsail instance from a stopped state}
\usage{
start_instance(instanceName)
}
\arguments{
\item{instanceName}{The name of the instance (a virtual private server) to start.}
}
\description{
Starts a specific Amazon Lightsail instance from a stopped state. To restart an instance, use the reboot instance operation.
}
\examples{

}
