% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{start_instance}
\alias{start_instance}
\title{Starts a specific Amazon Lightsail instance from a stopped state}
\usage{
start_instance(
  instanceName = "string"
)
}
\arguments{
\item{instanceName}{[required] The name of the instance (a virtual private server) to start.}
}
\description{
Starts a specific Amazon Lightsail instance from a stopped state. To restart an instance, use the reboot instance operation.
}
\details{
The \code{start instance} operation supports tag-based access control via resource tags applied to the resource identified by instanceName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\examples{

}
