% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{update_domain_entry}
\alias{update_domain_entry}
\title{Updates a domain recordset after it is created}
\usage{
update_domain_entry(domainName, domainEntry)
}
\arguments{
\item{domainName}{[required] The name of the domain recordset to update.}

\item{domainEntry}{[required] An array of key-value pairs containing information about the domain entry.}
}
\description{
Updates a domain recordset after it is created.
}
\details{
The \code{update domain entry} operation supports tag-based access control via resource tags applied to the resource identified by domainName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\section{Accepted Parameters}{
\preformatted{update_domain_entry(
  domainName = "string",
  domainEntry = list(
    id = "string",
    name = "string",
    target = "string",
    isAlias = TRUE|FALSE,
    type = "string",
    options = list(
      "string"
    )
  )
)
}
}

