% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lightsail_operations.R
\name{update_domain_entry}
\alias{update_domain_entry}
\title{Updates a domain recordset after it is created}
\usage{
update_domain_entry(domainName, domainEntry)
}
\arguments{
\item{domainName}{The name of the domain recordset to update.}

\item{domainEntry}{An array of key-value pairs containing information about the domain entry.}
}
\description{
Updates a domain recordset after it is created.
}
\examples{

}
